Conflicts:
	scintilla/Document.cxx
	scintilla/Editor.cxx
	scintilla/Editor.h
	scintilla/LineMarker.cxx
	scintilla/Style.cxx
	scintilla/Style.h
	scintilla/ViewStyle.h
	scintilla/XPM.cxx
	src/ScintillaBase.cxx

+++ scintilla/PositionCache.cxx
@@ -189,10 +189,10 @@ void LineLayout::RestoreBracesHighlight(Range rangeLine, Position braces[], bool
 	xHighlightGuide = 0;
 }
 
-int LineLayout::FindBefore(int x, int lower, int upper) const {
+int LineLayout::FindBefore(float x, int lower, int upper) const {
 	do {
 		int middle = (upper + lower + 1) / 2; 	// Round high
-		int posMiddle = positions[middle];
+		float posMiddle = positions[middle];
 		if (x < posMiddle) {
 			upper = middle - 1;
 		} else {
